requirements.txt
mysql-connector-python==8.1.0
tkcalendar==1.6.1
